+++ packages/google/src/RPA/Cloud/Google/keywords/apps_script.py
@@ -1,12 +1,9 @@
 from typing import Optional
 
-from . import (
-    LibraryContext,
-    keyword,
-)
+from . import keyword
 
 
-class AppsScriptKeywords(LibraryContext):
+class AppsScriptKeywords:
     """Class for Google Apps Script API
 
     For more information about Google Apps Script API link_.
@@ -15,8 +12,8 @@ class AppsScriptKeywords(LibraryContext):
     """
 
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.service = None
+        self.ctx = ctx
+        self.script_service = None
 
     @keyword(tags=["init", "apps script"])
     def init_apps_script(
@@ -38,7 +35,7 @@ class AppsScriptKeywords(LibraryContext):
         apps_scopes = ["script.projects", "drive.scripts", "script.external_request"]
         if scopes:
             apps_scopes += scopes
-        self.service = self.init_service(
+        self.script_service = self.ctx.init_service(
             service_name="script",
             api_version="v1",
             scopes=apps_scopes,
@@ -47,7 +44,7 @@ class AppsScriptKeywords(LibraryContext):
             use_robocorp_vault=use_robocorp_vault,
             token_file=token_file,
         )
-        return self.service
+        return self.script_service
 
     @keyword(tags=["apps script"])
     def run_script(
@@ -74,7 +71,7 @@ class AppsScriptKeywords(LibraryContext):
         if parameters:
             request["parameters"] = [parameters]
         response = (
-            self.service.scripts()
+            self.script_service.scripts()
             .run(
                 body=request,
                 scriptId=script_id,