+++ packages/google/src/RPA/Cloud/Google/keywords/drive.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Optional
 from apiclient.errors import HttpError
 from apiclient.http import MediaFileUpload, MediaIoBaseDownload
 
-from . import LibraryContext, keyword, UpdateAction
+from . import keyword, UpdateAction
 from .enums import DriveRole, DriveType, to_drive_role, to_drive_type
 
 
@@ -16,7 +16,7 @@ class GoogleDriveError(Exception):
     """Raised with errors in Drive API"""
 
 
-class DriveKeywords(LibraryContext):
+class DriveKeywords:
     """Class for Google Drive API
 
     For more information about Google Drive API link_.
@@ -25,8 +25,8 @@ class DriveKeywords(LibraryContext):
     """
 
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.service = None
+        self.ctx = ctx
+        self.drive_service = None
 
     @keyword(tags=["init", "drive"])
     def init_drive(
@@ -54,7 +54,7 @@ class DriveKeywords(LibraryContext):
         ]
         if scopes:
             drive_scopes += scopes
-        self.service = self.init_service(
+        self.drive_service = self.ctx.init_service(
             service_name="drive",
             api_version="v3",
             scopes=drive_scopes,
@@ -63,7 +63,7 @@ class DriveKeywords(LibraryContext):
             use_robocorp_vault=use_robocorp_vault,
             token_file=token_file,
         )
-        return self.service
+        return self.drive_service
 
     @keyword(tags=["drive"])
     def upload_drive_file(
@@ -107,7 +107,7 @@ class DriveKeywords(LibraryContext):
             raise GoogleDriveError("Filename '%s' does not exist" % filename)
 
         query_string = f"name = '{filepath.name}' and '{folder_id}' in parents"
-        self.logger.debug("Upload query_string: '%s'" % query_string)
+        self.ctx.logger.debug("Upload query_string: '%s'" % query_string)
         target_file = self.search_drive_files(query=query_string, recurse=True)
         guess_mimetype = mimetypes.guess_type(str(filepath.absolute()))
         file_mimetype = guess_mimetype[0] if guess_mimetype else "*/*"
@@ -119,15 +119,15 @@ class DriveKeywords(LibraryContext):
             "parents": [folder_id],
             "mimeType": file_mimetype,
         }
-        self.logger.debug("Upload file_metadata: '%s'" % file_metadata)
+        self.ctx.logger.debug("Upload file_metadata: '%s'" % file_metadata)
         if len(target_file) == 1 and overwrite:
-            self.logger.info("Overwriting file '%s' with new content", filename)
+            self.ctx.logger.info("Overwriting file '%s' with new content", filename)
             return self._file_update(target_file, media)
         elif len(target_file) == 1 and not overwrite:
-            self.logger.warn("Not uploading new copy of file '%s'", filepath.name)
+            self.ctx.logger.warn("Not uploading new copy of file '%s'", filepath.name)
             return target_file[0]["id"]
         elif len(target_file) > 1:
-            self.logger.warn(
+            self.ctx.logger.warn(
                 "Drive already contains '%s' copies of file '%s'. Not uploading again."
                 % (len(target_file), filepath.name)
             )
@@ -138,7 +138,7 @@ class DriveKeywords(LibraryContext):
     def _file_create(self, file_metadata, media):
         try:
             result = (
-                self.service.files()
+                self.drive_service.files()
                 .create(
                     body=file_metadata,
                     media_body=media,
@@ -153,7 +153,7 @@ class DriveKeywords(LibraryContext):
     def _file_update(self, target_file, media):
         try:
             result = (
-                self.service.files()
+                self.drive_service.files()
                 .update(fileId=target_file[0]["id"], media_body=media, fields="id")
                 .execute()
             )
@@ -163,7 +163,7 @@ class DriveKeywords(LibraryContext):
 
     def _download_with_fileobject(self, file_object):
         try:
-            request = self.service.files().get_media(fileId=file_object["id"])
+            request = self.drive_service.files().get_media(fileId=file_object["id"])
         except HttpError as err:
             raise GoogleDriveError(str(err)) from err
         fh = BytesIO()
@@ -221,12 +221,12 @@ class DriveKeywords(LibraryContext):
                 current_time = time.time()
                 files_downloaded.append(f["name"])
                 if limit and len(files_downloaded) >= limit:
-                    self.logger.info(
+                    self.ctx.logger.info(
                         "Drive download limit %s reached. Stopping the download.", limit
                     )
                     break
                 if timeout and (current_time - start_time) > float(timeout):
-                    self.logger.info(
+                    self.ctx.logger.info(
                         "Drive download timeout %s seconds reached. "
                         "Stopping the download.",
                         timeout,
@@ -323,10 +323,10 @@ class DriveKeywords(LibraryContext):
         else:
             # TODO: mypy should handle enum exhaustivity validation
             raise ValueError(f"Unsupported update action: {action}")
-        self.logger.debug(body)
+        self.ctx.logger.debug(body)
         try:
             updated_file = (
-                self.service.files().update(fileId=file_id, body=body).execute()
+                self.drive_service.files().update(fileId=file_id, body=body).execute()
             )
         except HttpError as err:
             raise GoogleDriveError(str(err)) from err
@@ -368,10 +368,10 @@ class DriveKeywords(LibraryContext):
         delete_count = 0
         for tf in target_files:
             try:
-                self.service.files().delete(fileId=tf).execute()
+                self.drive_service.files().delete(fileId=tf).execute()
             except HttpError as err:
                 if suppress_errors:
-                    self.logger.warn(str(err))
+                    self.ctx.logger.warn(str(err))
                 else:
                     raise GoogleDriveError(str(err)) from err
             delete_count += 1
@@ -403,7 +403,7 @@ class DriveKeywords(LibraryContext):
         if folder is None:
             try:
                 drive_file = (
-                    self.service.files().get(fileId="root", fields="id").execute()
+                    self.drive_service.files().get(fileId="root", fields="id").execute()
                 )
             except HttpError as err:
                 raise GoogleDriveError(str(err)) from err
@@ -416,7 +416,7 @@ class DriveKeywords(LibraryContext):
             if len(folders) == 1:
                 drive_file = folders[0]  # .get("id", None)
             else:
-                self.logger.info(
+                self.ctx.logger.info(
                     "Found %s directories with name '%s'" % (len(folders), folder)
                 )
         if drive_file:
@@ -470,11 +470,15 @@ class DriveKeywords(LibraryContext):
                 "Unable to find target folder: '%s'" % (target if target else "root")
             )
         for tf in target_files:
-            file = self.service.files().get(fileId=tf["id"], fields="parents").execute()
+            file = (
+                self.drive_service.files()
+                .get(fileId=tf["id"], fields="parents")
+                .execute()
+            )
             previous_parents = ",".join(file.get("parents"))
             try:
                 result_file = (
-                    self.service.files()
+                    self.drive_service.files()
                     .update(
                         fileId=tf["id"],
                         addParents=target_parent,
@@ -542,8 +546,8 @@ class DriveKeywords(LibraryContext):
             if page_token:
                 parameters["pageToken"] = page_token
             try:
-                self.logger.debug("Searching with parameters: '%s'" % parameters)
-                response = self.service.files().list(**parameters).execute()
+                self.ctx.logger.debug("Searching with parameters: '%s'" % parameters)
+                response = self.drive_service.files().list(**parameters).execute()
                 for file_details in response.get("files", []):
                     file_dict = self._drive_file_details_into_file_dict(file_details)
                     filelist.append(file_dict)
@@ -581,13 +585,13 @@ class DriveKeywords(LibraryContext):
         kind = details.get("kind")
         mimetype = details.get("mimeType")
         is_folder = mimetype == "application/vnd.google-apps.folder"
+        # fmt: off
         folder_id = (
             file_id
             if mimetype == "application/vnd.google-apps.folder"
-            else parents[0]
-            if parents and len(parents) > 0
-            else None
+            else parents[0] if parents and len(parents) > 0 else None
         )
+        # fmt: on
         file_link = (
             None
             if mimetype == "application/vnd.google-apps.folder"
@@ -642,7 +646,7 @@ class DriveKeywords(LibraryContext):
 
         folder_id = self.get_drive_folder_id(folder, parent_folder=parent_folder)
         if folder_id:
-            self.logger.info(
+            self.ctx.logger.info(
                 "Folder '%s' already exists. Not creating new one.", folder_id
             )
             return self._folder_response(folder_id)
@@ -657,7 +661,9 @@ class DriveKeywords(LibraryContext):
             file_metadata["parents"] = [parent_folder_id]
         try:
             added_folder = (
-                self.service.files().create(body=file_metadata, fields="id").execute()
+                self.drive_service.files()
+                .create(body=file_metadata, fields="id")
+                .execute()
             )
             return self._folder_response(added_folder["id"])
         except HttpError as err:
@@ -700,7 +706,7 @@ class DriveKeywords(LibraryContext):
         if len(target_files) != 1:
             raise ValueError("Did not find the Google Drive file to export")
         try:
-            request = self.service.files().export(
+            request = self.drive_service.files().export(
                 fileId=target_files[0], mimeType=mimetype
             )
         except HttpError as err:
@@ -812,7 +818,9 @@ class DriveKeywords(LibraryContext):
             request_parameters["emailMessage"] = notification_message
 
         try:
-            response = self.service.permissions().create(**request_parameters).execute()
+            response = (
+                self.drive_service.permissions().create(**request_parameters).execute()
+            )
             return {"file_id": target_file[0], "permission_id": response["id"]}
         except HttpError as err:
             raise GoogleDriveError(str(err)) from err
@@ -861,20 +869,20 @@ class DriveKeywords(LibraryContext):
         if not target_file:
             raise GoogleDriveError("Did not find target file")
 
-        self.logger.info(
+        self.ctx.logger.info(
             "Removing permission id '%s' for file_id '%s'"
             % (permission_id, target_file[0])
         )
         response = None
         try:
             response = (
-                self.service.permissions()
+                self.drive_service.permissions()
                 .delete(fileId=target_file[0], permissionId=permission_id)
                 .execute()
             )
         except HttpError as err:
             if suppress_errors:
-                self.logger.warn(str(err))
+                self.ctx.logger.warn(str(err))
             else:
                 raise GoogleDriveError(str(err)) from err
         return response
@@ -939,7 +947,7 @@ class DriveKeywords(LibraryContext):
         for tf in target_files:
             file_permissions_removed = []
             if "permissions" in tf and tf["permissions"]:
-                self.logger.info(
+                self.ctx.logger.info(
                     "Removing shares from file '%s' id '%s'" % (tf["name"], tf["id"])
                 )
                 for p in tf["permissions"]:
@@ -976,13 +984,13 @@ class DriveKeywords(LibraryContext):
         self, drive_file, permission, permissions_removed, suppress_errors
     ):
         try:
-            self.service.permissions().delete(
+            self.drive_service.permissions().delete(
                 fileId=drive_file["id"], permissionId=permission["id"]
             ).execute()
             permissions_removed.append(permission)
         except HttpError as err:
             if suppress_errors:
-                self.logger.warn(str(err))
+                self.ctx.logger.warn(str(err))
             else:
                 raise GoogleDriveError(str(err)) from err
 
@@ -1016,7 +1024,7 @@ class DriveKeywords(LibraryContext):
         permissions_removed = []
         for tf in target_files:
             if "permissions" in tf and tf["permissions"]:
-                self.logger.info(
+                self.ctx.logger.info(
                     "Removing shares from file '%s' id '%s'" % (tf["name"], tf["id"])
                 )
                 for p in tf["permissions"]:
@@ -1047,12 +1055,12 @@ class DriveKeywords(LibraryContext):
         response = None
         try:
             raw_response = (
-                self.service.files().get(fileId=file_id, fields="*").execute()
+                self.drive_service.files().get(fileId=file_id, fields="*").execute()
             )
             response = self._drive_file_details_into_file_dict(raw_response)
         except HttpError as err:
             if suppress_errors:
-                self.logger.warn(str(err))
+                self.ctx.logger.warn(str(err))
             else:
                 raise GoogleDriveError(str(err)) from err
         return response