+++ packages/google/src/RPA/Cloud/Google/keywords/speech_to_text.py
@@ -2,10 +2,7 @@ from typing import Dict, Optional
 from google.cloud import speech
 from google.cloud.speech_v1.types import RecognitionConfig, RecognitionAudio
 
-from . import (
-    LibraryContext,
-    keyword,
-)
+from . import keyword
 
 ENCODING = {
     "AMR": RecognitionConfig.AudioEncoding.AMR,
@@ -19,7 +16,7 @@ ENCODING = {
 }
 
 
-class SpeechToTextKeywords(LibraryContext):
+class SpeechToTextKeywords:
     """Class for Google Cloud Speech-To-Text API
 
     Possible input audio encodings:
@@ -39,8 +36,8 @@ class SpeechToTextKeywords(LibraryContext):
     """
 
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.service = None
+        self.ctx = ctx
+        self.speech_service = None
 
     @keyword(tags=["init", "speech to text"])
     def init_speech_to_text(
@@ -55,10 +52,10 @@ class SpeechToTextKeywords(LibraryContext):
         :param use_robocorp_vault: use credentials in `Robocorp Vault`
         :param token_file: file path to token file
         """
-        self.service = self.init_service_with_object(
+        self.speech_service = self.ctx.init_service_with_object(
             speech.SpeechClient, service_account, use_robocorp_vault, token_file
         )
-        return self.service
+        return self.speech_service
 
     @keyword(tags=["speech to text"])
     def recognize_text_from_audio(
@@ -90,7 +87,7 @@ class SpeechToTextKeywords(LibraryContext):
         """
         audio = self.set_audio_type(audio_file, audio_uri)
         parameters = {"use_enhanced": True}
-        # audio_encoding = ENCODING["UNSPECIFIED"]
+        parameters["encoding"] = ENCODING["UNSPECIFIED"]
         if encoding and encoding.upper() in ENCODING.keys():
             parameters["encoding"] = ENCODING[encoding.upper()]
         if sample_rate:
@@ -100,7 +97,7 @@ class SpeechToTextKeywords(LibraryContext):
         if audio_channel_count:
             parameters["audio_channel_count"] = audio_channel_count
         config = RecognitionConfig(**parameters)  # pylint: disable=E1101
-        rec = self.service.recognize(config=config, audio=audio)
+        rec = self.speech_service.recognize(config=config, audio=audio)
         return rec.results
 
     def set_audio_type(self, audio_file, audio_uri):