+++ packages/google/src/RPA/Cloud/Google/keywords/storage.py
@@ -2,13 +2,10 @@ from typing import Any, Dict, List, Optional
 
 from google.cloud import storage
 
-from . import (
-    LibraryContext,
-    keyword,
-)
+from . import keyword
 
 
-class StorageKeywords(LibraryContext):
+class StorageKeywords:
     """Class for Google Cloud Storage API
      and Google Cloud Storage JSON API
 
@@ -18,8 +15,8 @@ class StorageKeywords(LibraryContext):
     """
 
     def __init__(self, ctx):
-        super().__init__(ctx)
-        self.service = None
+        self.ctx = ctx
+        self.storage_service = None
 
     @keyword(tags=["init", "storage"])
     def init_storage(
@@ -34,10 +31,10 @@ class StorageKeywords(LibraryContext):
         :param use_robocorp_vault: use credentials in `Robocorp Vault`
         :param token_file: file path to token file
         """
-        self.service = self.init_service_with_object(
+        self.storage_service = self.ctx.init_service_with_object(
             storage.Client, service_account, use_robocorp_vault, token_file
         )
-        return self.service
+        return self.storage_service
 
     @keyword(tags=["storage"])
     def create_storage_bucket(self, bucket_name: str) -> Dict:
@@ -54,7 +51,7 @@ class StorageKeywords(LibraryContext):
 
             ${result}=   Create Storage Bucket   visionfolder
         """
-        bucket = self.service.create_bucket(bucket_name)
+        bucket = self.storage_service.create_bucket(bucket_name)
         return bucket
 
     @keyword(tags=["storage"])
@@ -94,7 +91,7 @@ class StorageKeywords(LibraryContext):
 
             ${result}=   Get Bucket   visionfolder
         """
-        bucket = self.service.get_bucket(bucket_name)
+        bucket = self.storage_service.get_bucket(bucket_name)
         return bucket
 
     @keyword(tags=["storage"])
@@ -114,7 +111,7 @@ class StorageKeywords(LibraryContext):
                 Log  ${bucket}
             END
         """
-        return list(self.service.list_buckets())
+        return list(self.storage_service.list_buckets())
 
     @keyword(tags=["storage"])
     def delete_storage_files(self, bucket_name: str, files: Any) -> List:
@@ -254,7 +251,7 @@ class StorageKeywords(LibraryContext):
                 if blob:
                     with open(filename, "wb") as f:
                         blob.download_to_file(f)
-                        self.logger.info(
+                        self.ctx.logger.info(
                             "Downloaded object %s from Google to filepath %s",
                             object_name,
                             filename,
@@ -268,7 +265,7 @@ class StorageKeywords(LibraryContext):
                 if blob:
                     with open(filename, "wb") as f:
                         blob.download_to_file(f)
-                        self.logger.info(
+                        self.ctx.logger.info(
                             "Downloaded object %s from Google to filepath %s",
                             filename,
                             filename,